backend/requirements.txt
requests==2.28.1
django==4.0.5
djangorestframework==3.13.1
django-filter==22.1
drf-extra-fields==3.4.0
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3    # если выдает ошибку
# pip3 install psycopg2==2.9.3
python-dotenv==0.20.0
Pillow==9.2.0